Key functionality and vowel layout
Added layout for vowels using a table that roughly mimics the IPA vowel chart (sometimes very roughly, but I think I've achieved a good balance between accuracy and complexity).

Additionally, worked out the basic typing functionality and added it to a single button. Next to do is make the other buttons, style them up nice, and make sure the code adding the event listener to all buttons is adding it to all the buttons.